[Refactor] LottoBase 삭제 후 Lotto 관련 클래스에 책임 분리
